loader: Avoid using "name" for the to-be-executed file in SimpleTest
Currently simpletest uses the "name" to define which file will be
executed as the binary. This does not look right and can break in case
one postprocesses the names in the testsuite. Let's introduce an
argument to define the name and (for now) fall-back to "name.name" in
case it's "None".

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>