test_mux.TestMuxTree.test_basic_functions: get a string on Python 3
And not bytes.  What happens here is that
`avocado.core.tree.tree_view()` will return either bytes with 'utf-8'
or 'ascii' depending on the given parameter.

On Python 2, it's fine to compare the result with strings.  On Python
3, we have to get a string from the bytes returned, that is, we need
to decode the bytes.

While at it, let's fix a typo (s/asci/ascii).

Signed-off-by: Cleber Rosa <crosa@redhat.com>