selftests/functional/test_output.py: use bytes in comparison
The content on the file is being read as bytes, so let's use byte
literals in the comparison.

The error message received some adjustments, mainly because it could
produce weird stuff text such as "1th" or "2th".  To be honest, I'd
remove this error message completely, the assertion error is pretty
easy to spot IMO.

Signed-off-by: Cleber Rosa <crosa@redhat.com>