Style only change: make it clear that there's a node being created
The current code style creates and appends an object into the list,
and then later implicitly references that item using the `[-1]`
notation, changing its filter.

While correct, IMO, it's easier to follow code that creates the
object, sets its values, and then appends it to the list.

This was a style only change that appeared while debugging another
mux related issue.

Signed-off-by: Cleber Rosa <crosa@redhat.com>