avocado/core/tree.py: properly copy the TreeEnvironment
On Python 3, trying to copy a FilterSet, using the instance.copy()
method, will return a plain set() object.  This influences a number of
code, including the fingerprint generation, and the variant ID which
will be based on the hash of the fingerprint.

Let's make things predictable and copy using Python's standard module
copy.

This fixes the functional test in module `test_getdata` under Python
3, as the variant IDs are now the same on jobs run under Python 2 and
3.

Signed-off-by: Cleber Rosa <crosa@redhat.com>