Avocado CLI App: always write crash file encoded in utf-8
When things go wrong, the Avocado command line app will write a file
with the backtrace that caused the trash.  Depending on its content,
it may not be possible to encode that in an file encoded as ASCII.
Let's default to utf-8 then.

Also, this is done for portability reasons, as it brings a consistent
behavior (one explicit encoding type) across Python 2 and 3, which
have different default encodings.

Signed-off-by: Cleber Rosa <crosa@redhat.com>