Packaging: update classifiers and install requirements
We are, finally, advertising Python 3 support.  Yes, it's supposed to
be "beta" quality level, as there are a few leftover issues we know
about, and possibly a whole lot more we don't.

Still, we're taking that leap of faith, and that suggests describing
that so that PyPI users and installation tools know about it.

Reference: https://packaging.python.org/tutorials/distributing-packages/#classifiers
Reference: https://packaging.python.org/tutorials/distributing-packages/#python-requires
Signed-off-by: Cleber Rosa <crosa@redhat.com>