RPM: use exact package names on requires
We currently build on EL7, Fedora 26 and Fedora 27.

On EL7, some of the Python packages are prefixed with "python2", and
some others are not.  We have been previously using the name of
"Provides" instead of the real package names.

Now, Python package names on Fedora 26 and Fedora 27 are the same for
those we depend on.

This adds a conditional block for EL7, and another one for Fedora 26
and 27.  Some Python packages have been kept on both blocks, in
preparation for the Python 3 packaging that will happen only for the
Fedora packages.

Signed-off-by: Cleber Rosa <crosa@redhat.com>