Test: provide more metadata information
The amount of data to log is not an exact science, and certainly a
matter of preference.  Having said that, we already have a section
about test metadata, and IMO, a few other test metadata can influence
the test results and ease debugging and troubleshooting.

With that in mind, let's log a few more of the test metadata.  It's
categorized as DEBUG messages anyway, and in IMO it even looks better
to have a section with more than one piece of information.

Signed-off-by: Cleber Rosa <crosa@redhat.com>