selftests/functional/test_json_variants.py: do not rely on the order of dict
The items of a dictionary are not ordered by default, and between
Python 2 and 3 they always come out different in my experience.

Since this test is trying out the representation of a JSON built from
dicts, the same applies.  Let's make the checks a bit more flexible
(and a bit less accurate to be honest) and check for the pieces of
info inside the JSON file content.

Signed-off-by: Cleber Rosa <crosa@redhat.com>