selftests/functional/test_thirdparty_bugs.py: feed string to json
As it was done in d2b4edc, Python 3.4 needs a string to be passed to
json.loads() (and, indirectly, to json.load()).  Different from
d2b4edc, the content we fetch externally may be able to tell us the
encoding that is used.

Let's use that information, if available, to decode the stream of
bytes into a string that is acceptable to json.loads().

Signed-off-by: Cleber Rosa <crosa@redhat.com>