selftests/unit/test_utils_process.py: remove decorators to skip tests
On the basis of the non-existence of the "true" command on the system
running tests.  Most of the tests do not even reference the "true"
command, and those that do, do not require the binary to exist, but
only a valid command line to be given.

In that aspect, a fictional command is used where needed.

Signed-off-by: Cleber Rosa <crosa@redhat.com>