avocado.utils.process: require commands to be given as unicode strings
This change introduces the following requirement: the command to
be run by the avocado.utils.process functions should be given as unicode.
It brings a clearer interface, in which there's no need to deal with
encodings internally.

For simple test, it brings changes to respect the of passing unicode.
And, it changes the presentation of simple tests with unicode names
explicitly expreseed as such.

The way to deal with the process data that may be generated on
stdout/stderr is unchanged.

Signed-off-by: Cleber Rosa <crosa@redhat.com>