[RFC] Enable W1618 and W0410 warnings on lint
This is sent as RFC, because I've felt that applying the
recommendations of W1618 (and related W0410) to example/tests/*, can
introduce yet another level of boilermuch code, adding noise, and
reducing their value.

Basically, a simple example like passtest.py, would start to look like:

------------------------------------------------------------------------------

  #!/usr/bin/env python

  from __future__ import absolute_import

  from avocado import main
  from avocado import Test

  class PassTest(Test):

      """
      Example test that passes.

      :avocado: tags=fast
      """

      def test(self):
          """
          A test simply doesn't have to fail in order to pass
          """
          pass

  if __name__ == "__main__":
      main()

------------------------------------------------------------------------------

Feedback is welcome, specially on if/how this affects the readability
and purpose of the example tests.

Signed-off-by: Cleber Rosa <crosa@redhat.com>