selftests/functional/test_basic.py: make echo output safe
In the module global scope, "man echo" is executed to attempt to
determine if it's a GNU echo.  But, "man echo" in a locale such
as UTF-8 produces funky copyright characters:

  $ LANG=en_US.UTF-8 man echo | chardetect -
  <stdin>: windows-1252 with confidence 0.73

Which can not be decoded by the UTF-8 codec, producing
UnicodeDecodeErrors.  Even if that wasn't the case, it's better to
control the locale of command that can generate varied output, and
that we depend on.

  $ LANG=C man echo | chardetect -
  <stdin>: ascii with confidence 1.0

Signed-off-by: Cleber Rosa <crosa@redhat.com>