FDDrainer: pass text to loggers
The FDDrainer reads from process output and write to loggers.  The
test log, which is intended for humans, can indeed try to treat it as
text.  The stream logger, on the other hand, should deal with data as
binary content *only*, but this will be handled on a future change.

By having text content passed to the loggers, we avoid the occurrence
of log lines that look like:

  [stdout] b'command output'

Or output files (job-YYYY-MM-DDTHH.MM-id/tests/<test>/output) that
may literally contain:

  b'command output'

The encoding used is the one specified at the command execution.  With
this fix, a number of tests on Python 3 can now run successfully.

Signed-off-by: Cleber Rosa <crosa@redhat.com>