selftests/unit/test_xunit.py: specify JUnit variation used
It's not easy to find one definitive specification for the the JUnit /
Xunit formats, and a number of different schemas have been published
by different entities.

The current schema version we ship was based on a repository that is
no longer available.  Even though the format is exactly the same,
let's switch to the schema hosted by the JUnit project, and to what
is considered the Jenkins schema (as part of the JUnit 5 project).

Reference: https://junit.org/junit5/
Reference: https://raw.githubusercontent.com/junit-team/junit5/master/platform-tests/src/test/resources/jenkins-junit.xsd
Signed-off-by: Cleber Rosa <crosa@redhat.com>