Replay with remote functional test: skip if remote plugin is not available
When a plugin is not installed, its command line options are not
registered with the optparse based code.  That is caught earlier than
Avocado's own handling of incompatible options (in this case, the replay
and the remote runner), and results in the errors such as:

  avocado run: error: unrecognized arguments: --remote-hostname

When building RPM packages for Python 3, the remote runner plugin is
not available, and this test should be skipped.

Signed-off-by: Cleber Rosa <crosa@redhat.com>