utils.process: Support non-default encoding
When non-system encoding is selected in avocado.utils.process, it breaks
quite badly. The biggest problem is we internally turn anything into
unicode, which is then executed by python's subprocess as unicode,
therefor with the command already spoilt.

This patch changes the cmd to be always encoded bytes (still accepting
decoded one and encoding it using provided encoding) which avoids the
command to be spoiled by python's subprocess.

For similar reasons it's now necessary to pass encoding to all helper
functions that deal with the command and as encoding matters the
`cmd_split` command now returns the encoded bytes object.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>