optional_plugins: Limit fabric version to <2.0.0
The fabric-2.0.0 is a complete rework of fabric using different
approach and will require many changes to our code. Let's keep using the
1.x.y version for now.

The fabric3 is still only based on 1.x.y branch but let's also limit
that one as it is based on the original fabric so it's expected that
they might eventually come-up with 2.x.y py3 fork.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>