loader: Discover tests as "recursive" by default
For a while we have the ":avocado: recursive" tag, which allows
recursive discovery of tests and should limit the need for ":avocado:
enable".

This is a simple version, which only changes the default and acts as if
every class was given the ":avocado: recursive" tag but we might want to
tweak the behavior of disable/enable/recursive in the near future to
improve the user experience.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>