CPU utils: handle /proc/* and /sys/* as binary
The Linux Kernel will not change the content of those files based on
locale.  So, let's stick with honest binary mode when reading those
files.

While at it, also change the regex patterns, as they wi'll be matched
agains /proc/* and /sys/* content, which should really be treated as
binary.

Signed-off-by: Cleber Rosa <crosa@redhat.com>
Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>