utils.astring: Add "to_bytes" to complement "to_text"
The "to_bytes" is useful when we need a representation of "anything"
encoded in bytes. This function is necessary because on py2 we can't
simply use "str(object)" as "object" might contain non-ascii characters
and on python3 we can't do "bytes(object)" because there are many
special variables (int => initialize number of \x00; [int, int] =>
create array of those numbers...).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>