plugins.tap: Flush only on "important" output
The TAP plugin is important source of information in case the machine
crashes, which is why we do force-flush on each logged line. Anyway this
is probably not that important when re-printing the test logs. Let's add
explicit "__flush" method and use it only on places where important info
(test progress) were printed.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>