utils.cpu: Avoid multiple-transformation of disable value
The legacy disable value is currently transformed in a loop using the
same variable, which results in TypeError as the transformed output is
not a valid input. Let's fix it by only single transformation in case
`setstate` is not used.

With `setstate` the value can be safely overridden as the value comes
from `setstate`.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>