Test Loader: use enums from from Python 3.4
This change reveleaed a couple of location in the symbolic names where
not being used, but their special values were being evaluated.  Let's
avoid using the special (and subtle) meaning of ALL (True) and
DEFAULT/AVAILABLE (which evaluate to False).

The enum library is part of the standard library, starting from 3.4
(which is the mininum version we require for Avocado).  On Python 2.7,
we can rely on the backport package.

This is of course not necessary right now, but I feel that we can
continue modernizing the code base, and not wait for 2020 and then be
liberated to use newer constructs and libraries.

Signed-off-by: Cleber Rosa <crosa@redhat.com>