Asset Fetcher: add a fetch/build example test
This example test fetches two different files from different
locations, and uses a data file (shipped *with* the test) on
one of them.

The goal is to demonstrate a complete example of files you
may want to distribute with your test, and the assets you
may want to fetch from remote locations.

Signed-off-by: Cleber Rosa <crosa@redhat.com>