avocado/utils/build.py: add a configure utility function
The most common workflow when building software from source is to
configure the source, and then build it.  Avocado provides a utility
"make" function, but when users need to run "configure" or similar
scripts, there's no shortcut.

This introduces a configure script, which when used with no explicit
configure script name, will attempt to find a couple of suitable
scripts, but will forgive the source tree if one doesn't exist.

This allows test writers to use the following pattern:

   from avocado.utils import build
   build.configure("/path/to/src")
   build.make("/path/to/src")

And have successful results in most source code repositories.

Signed-off-by: Cleber Rosa <crosa@redhat.com>