Asset Fetcher: remove hashfile attribute
In another part of the purge of attributes, let's remove hashfile.
This attribute is yet another case of an attribute that is not
intended for end users, and basically serves as a way to pass
information across methods.

Because the hash file name is so easily computable, instead of turning
it into a private, let's just evaluate the hash file name when needed.

Signed-off-by: Cleber Rosa <crosa@redhat.com>