[DO NOT MERGE] Travis CI tests
This is a collection of commits:

------------------------------------------------------------------------------

Remote runner: provide fabric version for Python 3 on non-Fedora

Fedora 29 (currently rawhide) has a python2-fabric3 package, so it can
use that (Fabric3) as a requirement on both Python 2.7 and 3.4+.  But
we have not switched to Fabric3 on other environments, so let's keep
the fabric (1.x) series there.

Signed-off-by: Cleber Rosa <crosa@redhat.com>

------------------------------------------------------------------------------

Travis CI: pin pylint version

Avocado uses inspekt, which in turn uses pylint.  On Travis-CI under
Python 2.7, pylint 1.9.2 is being used, while on Python 3.4 it's
installing 2.0.0.

The pylint 2.0.0 API is different, and produces the following error:

   Running 'inspekt lint --exclude=.git --enable W0101,W0102,W0404,W0611,W0612,W0622'
   Pylint disabled: W,R,C,E1002,E1101,E1103,E1120,F0401,I0011
   Pylint enabled : W0101,W0102,W0404,W0611,W0612,W0622
   __init__() got an unexpected keyword argument 'exit'
   lint FAILED

Let's pin the pylint version for predictable results, as we do with
pycodestyle.

Reference: https://trello.com/c/LDbYJFeR/1374-lint-failed-on-python-3
Signed-off-by: Cleber Rosa <crosa@redhat.com>

------------------------------------------------------------------------------

selftests/functional/test_output_check.py: fix Python 3.4 syntax

Formatting bytes was only introduced on Python 3.5, so let's just
add the bytes. At the same time, there's no need for some of that
content to live in the module namespace, as only STDOUT and STDERR
are used later in the tests.

Signed-off-by: Cleber Rosa <crosa@redhat.com>