Travis-CI: remove "system_site_packages" option
The "system_site_packages" option is needed when the Python
installation on the virtual environments need to access Python
packages outside of them, that is, on the system.  Given the
environment in which Travis-CI jobs run, this means packages installed
via apt, and not via pip.

While we do install packages via apt, we install libvirt-dev, and
install libvirt-python via pip.  In summary, we do not need that
option.

Signed-off-by: Cleber Rosa <crosa@redhat.com>