Avocado Instrumented Tests: move from loader to safeloader
Given that the implementation of the code to find Avocado Instrumented
Tests is already quite big, and that a good part of it already lives
in the safeloader module, let's complete the move.

This sets the tone of having the generic loader mechanism in the
"avocado.core.loader" module, while having implementations, at least
for the instrumented test at this point, in the "avocado.core.safeloader"
module.

With this change, the loader knows nothing about ast, and relies on
safeloader.find_avocado_tests() only.

Signed-off-by: Cleber Rosa <crosa@redhat.com>