avocado.utils.vmimage: allow users to specify a snapshot dir
Which is used for the (usually temporary) snapshot image.  By default,
it's kept in the on the same directory as the base image.  When
downloading the base image into a test's cache_dir, it makes sense to
keep the snapshot out of the cache_dir, usually in the test's workdir.

Signed-off-by: Cleber Rosa <crosa@redhat.com>