optional_plugins.varianter_cit: Fix the "path" value
The "paths" is supposed to be a list of paths and not just string.
Fortunatelly this was why this plugin actually worked, because it turned
paths into `['/', 'r', 'u', 'n']` making the root node available. Anyway
let's fix the value and use the only location this plugin puts variables
in, which is `/`.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>