Fix E1111 pylint error
E1111 error happens when a function which doesn't return is used to
assign to something. In this particular case the unittest.TextTestResut
stopTest method doesn't return and its result is being assigned to a
variable. This behaviour makes pylint fails with E1111 error. This
commit just don't assign the result of super() call and also keeps the
super class method behavior (not returing).

Signed-off-by: Caio Carrara <ccarrara@redhat.com>