avocado.utils.vmimage: operate on unique versions found
Sometimes multiple links to the same OS image file version will exist
on an index web page.  For instance, on openSUSE download server,
a folder icon, and the name of the version itself, appear side
by side and point to the same version:

               Name                    Last Modified  Size  Metadata
 [parent_icon] Parent Directory                       -
 [folder_icon] Leap_15.0/              08-Feb-2018    -
 ...

Let's make the set of found patterns unique, and not one per link.

Signed-off-by: Cleber Rosa <crosa@redhat.com>