avocado/utils/download.py: log retrieved URL information
Sometimes the URL used is a link to the content that gets updated
between jobs.  When comparing jobs, it's unclear if we're downloading
the same file or another one.

Let's print the available information we have, such as the content
length, and the creation/modification dates.

Reference: https://trello.com/c/wQ8ir7xA
Signed-off-by: Cleber Rosa <crosa@redhat.com>