avocado/utils/partition.py: do not blow up on systems with no lsof
lsof is not an Avocado dependency, and is much harder to find on
base systems than mount, kill, etc.  Since this module pretty
much revolves around those utilities, I'm not going to propose
major changes, trying to adapt to the lack of mount, kill, etc.

This simply does *not* attempt to run kill, if we can not get
the PIDs using lsof.  Then, the forceful unmount will probably
fail just the same, but the right way.

To be completely honest, there are still a broken expectation here:
when cannot assume that by sending a signal the process will be
killed.  The real is a lot more complex, though.

https://trello.com/c/DDUNPCBx
Signed-off-by: Cleber Rosa <crosa@redhat.com>