avocado/utils/partition.py: do not blow up on systems with no lsof
lsof is not an Avocado dependency, and is much harder to find on base
systems than mount, kill, etc.  Since this module pretty much revolves
around those utilities this does not propose major changes, trying to
adapt to the lack of mount, kill, etc.

This change makes it clear, through a PartitionError, when it's not
possible to run a forceful unmount due to the lack (or error) of lsof.
While at it, it refactors the methods that returns the list of PIDs
using lsof.

To be completely honest, there is still a broken expectation here: one
cannot assume that by sending a signal the process will be killed.
But, at least it is signalled by another PartitionError exception.

https://trello.com/c/DDUNPCBx
Signed-off-by: Cleber Rosa <crosa@redhat.com>