settings: Change the way config files are processed
Originally we only processed files that existed and when they were read
we updated list of config locations and in case of no return list of
failed-to-read paths. Anyway the "config.read" function accepts list of
config files and returns list of successfully parsed files. Missing
files are automatically skipped and broken files raise exception, so
let's just simply create list of to-be-read files and feed it into the
"config.read" function.

As a result the list of not-read files actually works, still in this
version it's only displayed on "avocado config", where users can get
notion of all possible locations and in "avocado run" only parsed files
are displayed as it doesn't seem important enough to log them there to
me.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>