Adds unit tests for utils.memory.get_buddy_info
The tests was done mocking the real output of `/proc/buddyinfo`. To make
it easier the process to get the buddy info content was extracted for a
specific function that actually only wraps the open call. This way we
can mock this function return only instead of try mock the builtin open
function.

Signed-off-by: Caio Carrara <ccarrara@redhat.com>