avocado.utils.vmimage: workaround implementation limitations for Fedora
This allows user to ask for a image for a "fedora" image on a "ppc64"
or "ppc64le" arch, without being exposed to the fact that Fedora hosts
those images in a different location (and treats them as secondary
arches).  Those images are still "Fedora", and that's what users need
to know.

The better fix here would be one that only determines the location of
the images (`url_versions`) based on the architecture.  It would be
possible to do this here, maybe, by using a property for
`url_versions` instead.  But this breaks the general design of the
code and maybe some expectations.

This change chooses to defer that to a possibly more coprehensive
change in the ImageProviderBase class, making the more dynamic
evaluation standard or common.

Signed-off-by: Cleber Rosa <crosa@redhat.com>