avocado.utils.vmimage: differentiate against base and concrete classes
First of all, a disclaimer: this is hackish.  But, it's the simplest
and thus best fix I could think of.

What happens here is that this module makes no distintcion between
what is a base class "abstract" class, and a concrete class.  With the
goal of reducing code duplication, a goal that will remain valid here
and in other modules, it made sense to create a common Fedora class.

Now, because only being a ImageProviderBase suffices, the
FedoraImageProviderBase class, which is not intended to be used as a
concrete (final) class, is being carried along and its "name" is
checked, causing crashes.

This introduces the ad-hoc, hackish, notion, that a concrete
ImageProvider is one that contains a name.

Signed-off-by: Cleber Rosa <crosa@redhat.com>