Improve dependency specification for tests and ci
The dependencies for selftests and CI execution was split in two
requirements files. It causes issues with libs being in different
versions when tests are executed locally and in CI environment. To
reduce theses issues and others the changes bellow was done.

All dependencies listed in requirements.txt are now specified only in
setup.py properly. This way we avoid duplications or missing
requirements that are necessary for avocado.

It also could be done keeping only requirements.txt file and loading it
in setup.py module. However, since avocado-framework have not clear use
case of requirements being installed via `pip install requirements.txt`,
it's better keep the dependencies only in setup module.

It was removed the requirements file only for CI purpose and
updates the requirements for selftest. This way we can have proper
dependencies installed when run selftests both in local or CI
environements.

The requirements-selftest was updated: Setuptools was removed since it's
already a dependency from avocado itself. Pep8 was removes since it's
already a dependecy from inspektor. Since we run lint checks by
inspektor it's better let it manage de pep8 (or pycodestyle) dependency

Signed-off-by: Caio Carrara <ccarrara@redhat.com>